create post

+++ app/Http/Controllers/TasksController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Task;    // add
+use App\Http\Controllers\Controller;
 
 class TasksController extends Controller
 {
@@ -15,11 +15,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-
-        return view("tasks.index", [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if(\Auth::check()){
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at','desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+                ];
+            $data += $this->counts($user);
+            return view('users.show',$data);
+        }else{
+            return view('welcome');
+        }
     }
 
     /**
@@ -43,17 +52,18 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
+
         $this->validate($request,[
             'content'=> 'required|max:191',
-            'status' => 'required|max:10',
+            'status' => 'required|max:191',
             ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            "content" => $request->content,
+            "status"  => $request->status,
+            ]);
 
-        return redirect('/');
+        return redirect()->back();
     }
 
     /**
@@ -97,7 +107,7 @@ class TasksController extends Controller
     {
         $this->validate($request,[
              'content' => 'required|max:191',
-             'status' => 'required|max:10',
+             'status' => 'required|max:191',
             ]);
         
         $task = Task::find($id);
@@ -116,9 +126,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
+        
+        if(\Auth::id() === $task->user_id){
+            $task->delete();
+        }
 
-        return redirect('/');
+        return redirect()->back();
     }
 }
